Another attempt to get the tilt motor break working.
Now we embed the command to disable the brake disable within
the axis command so it get executed in correct order.  specifically

SH D : enable the motor power to tilt
SB 9 : enable output to relay, disabling the brake for tilt
BG D : start move

Unfortunately I needed to hardcode this change for a particular axis;
this means this code will NOT work for gantry 1.